+++ sdk/recoveryservices-backup/Azure.ResourceManager.RecoveryServicesBackup/tests/ScenarioTests/BackupProtectionContainerTests.cs
@@ -13,6 +13,7 @@ using NUnit.Framework;
 
 namespace Azure.ResourceManager.RecoveryServicesBackup.Tests
 {
+    [NonParallelizable]
     public class BackupProtectionContainerTests : RecoveryServicesBackupManagementTestBase
     {
         public BackupProtectionContainerTests(bool isAsnyc)
@@ -20,7 +21,7 @@ namespace Azure.ResourceManager.RecoveryServicesBackup.Tests
         {
         }
 
-        [Test]
+        [RecordedTest]
         public async Task CreateTest()
         {
             var sub = await Client.GetDefaultSubscriptionAsync();