+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.0.0-preview.42 (Unreleased)
+## 12.0.0-preview.43 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,9 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.0.0-preview.42 (2023-12-05)
+- Added support for service version 2024-02-04.
 
 ## 12.0.0-preview.41 (2023-11-13)
 - Distributed tracing with `ActivitySource` is stable and no longer requires the [Experimental feature-flag](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/core/Azure.Core/samples/Diagnostics.md).